Misc improvements (#21)
* Get rid of a suspicious fs package

* Get rid of children rendering in TestshotComponent

* Rework building process

The current approach to the library build is superfluous and
inconvenient for the development. Now, webpack produces
an uglified single file with react, html-differ, etc. bundled in.

Instead of that I’ve updated build script to use Babel and compile
only ES & JSX syntax, but keep `require` calls as is.

* Remove post-install script

There are no need to build the library on every install. This script
is supposed to improve dev experience (as you don't need to build it
manually) but in fact to slow things down as it runs every time
`npm install` is ran.

* Fix entry point

- Add the Testshot component to the entry point
- Get rid of absolute paths

* Get rid of unused dependencies

* Allow to pass custom className to the Testshot container

- Get rid of rigid `z-index`
- Allow to pass a custom className for the Testshot container

* Use createElement to build a scenario element

- Use `React.createElement` instead of a plain function call
  so that it works with class components.

* Commit updated build